LPS-59825 rename packages
