+++ modules/apps/map/map-api/bnd.bnd
@@ -1,5 +1,5 @@
-Bundle-Name: Liferay Frontend Map API
-Bundle-SymbolicName: com.liferay.frontend.map.api
+Bundle-Name: Liferay Map API
+Bundle-SymbolicName: com.liferay.map.api
 Bundle-Version: 1.0.0
-Export-Package: com.liferay.frontend.map.api.*
+Export-Package: com.liferay.map.api.*
 Include-Resource: classes
