+++ modules/apps/map/map-api/src/main/java/com/liferay/map/api/BaseJSPMapProvider.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.frontend.map.api;
+package com.liferay.map.api;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
