+++ modules/apps/map/map-base-web/bnd.bnd
@@ -1,6 +1,6 @@
-Bundle-Name: Liferay Frontend Map Base Web
-Bundle-SymbolicName: com.liferay.frontend.map.base.web
+Bundle-Name: Liferay Map Base Web
+Bundle-SymbolicName: com.liferay.map.base.web
 Bundle-Version: 1.0.0
 Include-Resource: classes
 JS-Config: /META-INF/resources/js/config.js
-Web-ContextPath: /frontend-map-base-web
+Web-ContextPath: /map-base-web
