+++ modules/apps/map/map-openstreet-web/bnd.bnd
@@ -1,6 +1,6 @@
-Bundle-Name: Liferay Frontend Map OpenStreet Web
-Bundle-SymbolicName: com.liferay.frontend.map.openstreet.web
+Bundle-Name: Liferay Map OpenStreet Web
+Bundle-SymbolicName: com.liferay.map.openstreet.web
 Bundle-Version: 1.0.0
 Include-Resource: classes
 JS-Config: /META-INF/resources/js/config.js
-Web-ContextPath: /frontend-map-openstreet-web
+Web-ContextPath: /map-openstreet-web
