+++ modules/apps/map/map-openstreet-web/src/main/java/com/liferay/map/openstreet/web/OpenstreetMapProvider.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.frontend.map.openstreet.web;
+package com.liferay.map.openstreet.web;
 
-import com.liferay.frontend.map.api.BaseJSPMapProvider;
-import com.liferay.frontend.map.api.MapProvider;
+import com.liferay.map.api.BaseJSPMapProvider;
+import com.liferay.map.api.MapProvider;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
@@ -63,7 +63,7 @@ public class OpenstreetMapProvider extends BaseJSPMapProvider {
 
 	@Override
 	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.frontend.map.openstreet.web)",
+		target = "(osgi.web.symbolicname=com.liferay.map.openstreet.web)",
 		unbind = "-"
 	)
 	public void setServletContext(ServletContext servletContext) {
