+++ modules/apps/map/map-taglib-freemarker-contributor/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: Liferay Frontend Map Taglib FreeMarker Contributor
+Bundle-Name: Liferay Map Taglib FreeMarker Contributor
 Bundle-SymbolicName: com.liferay.map.taglib.freemarker.contributor
 Fragment-Host: com.liferay.portal.template.freemarker
 Bundle-Version: 1.0.0
