+++ modules/apps/map/map-taglib/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: Liferay Frontend Map Taglib
+Bundle-Name: Liferay Map Taglib
 Bundle-SymbolicName: com.liferay.map.taglib
 Bundle-Version: 1.0.0
 Export-Package:\
@@ -8,4 +8,4 @@ Include-Resource:\
 	META-INF/resources=src/main/resources/META-INF/resources
 Provide-Capability:\
 	osgi.extender;osgi.extender="jsp.taglib";uri="http://liferay.com/tld/map";version:Version="1.0.0"
-Web-ContextPath: /frontend-map-taglib
+Web-ContextPath: /map-taglib
