+++ modules/apps/map/map-taglib/src/main/resources/META-INF/resources/init.jsp
@@ -23,7 +23,7 @@ taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %><%@
 taglib uri="http://liferay.com/tld/util" prefix="liferay-util" %>
 
-<%@ page import="com.liferay.frontend.map.api.MapProvider" %><%@
+<%@ page import="com.liferay.map.api.MapProvider" %><%@
 page import="com.liferay.portal.kernel.language.LanguageUtil" %><%@
 page import="com.liferay.portal.kernel.util.GetterUtil" %><%@
 page import="com.liferay.portal.kernel.util.ResourceBundleUtil" %><%@
