+++ modules/apps/map/map-taglib/src/main/resources/META-INF/resources/map_display/init.jsp
